test(mapgen): align tests to causal owners
