+++ mods/mod-swooper-maps/test/domains/resources/sites/selection.test.ts
@@ -4,7 +4,6 @@ import { admitPositiveResourceRegionMinimum } from "@mapgen/domain/resources";
 import resources from "@mapgen/domain/resources/router";
 import { hexDistanceOddQPeriodicX } from "@swooper/mapgen-core/lib/grid";
 import { runAdmittedOperationForTest } from "@swooper/mapgen-core/testing";
-import { artifacts as resourceSiteArtifacts } from "@mapgen/domain/resources/modules/sites/artifacts/index.js";
 
 type SelectInput = Parameters<typeof resources.sites.ops.selectResourceSites.run>[0];
 type Demand = Pick<
@@ -225,57 +224,6 @@ describe("select-resource-sites operation contract", () => {
         count: 4,
       },
     ]);
-  });
-
-  it("rejects stale terminal deficit evidence at the resource-plan artifact boundary", () => {
-    const syntheticDimensions = { width: 4, height: 3 } as const;
-    const input = buildInput({
-      ...syntheticDimensions,
-      demands: [
-        {
-          resourceType: "RESOURCE_A",
-          weight: 10,
-          targetCount: 8,
-          minCount: 1,
-          maxCount: 10,
-        },
-      ],
-    });
-    const result = run(input);
-    const context = { dimensions: syntheticDimensions };
-    expect(resourceSiteArtifacts.resourcePlan.validate(result, context)).toEqual([]);
-
-    const missing = structuredClone(result);
-    missing.perType[0]!.shortfalls.splice(0);
-    expect(
-      resourceSiteArtifacts.resourcePlan
-        .validate(missing, context)
-        .some((entry) => entry.message.includes("requires one terminal shortfall"))
-    ).toBe(true);
-
-    const wrongType = structuredClone(result);
-    wrongType.perType[0]!.shortfalls[0]!.resourceType = "RESOURCE_B";
-    expect(
-      resourceSiteArtifacts.resourcePlan
-        .validate(wrongType, context)
-        .some((entry) => entry.message.includes("names another resource type"))
-    ).toBe(true);
-
-    const wrongCount = structuredClone(result);
-    wrongCount.perType[0]!.shortfalls[0]!.count += 1;
-    expect(
-      resourceSiteArtifacts.resourcePlan
-        .validate(wrongCount, context)
-        .some((entry) => entry.message.includes("terminal deficit"))
-    ).toBe(true);
-
-    const stale = structuredClone(result);
-    stale.perType[0]!.effectiveTargetCount = stale.perType[0]!.plannedCount;
-    expect(
-      resourceSiteArtifacts.resourcePlan
-        .validate(stale, context)
-        .some((entry) => entry.message.includes("requires no terminal shortfall"))
-    ).toBe(true);
   });
 
   it("runs the region-minimum force pass only for an admitted required state (E2.2)", () => {
