+++ mods/mod-swooper-maps/test/recipes/swooper-physics-standard/stages/placement/placement-product-reconciliation.test.ts
@@ -1,18 +1,11 @@
 import { describe, expect, it } from "bun:test";
 
-import {
-  createMockAdapter,
-  MockAdapter,
-  type OfficialDiscoveryGenerationResult,
-  type ResourcePlacementIntent,
-  type ResourcePlacementOutcome,
-} from "@civ7/adapter";
+import { MockAdapter, type OfficialDiscoveryGenerationResult } from "@civ7/adapter";
+import { artifacts as resourceSiteArtifacts } from "@mapgen/domain/resources/modules/sites/artifacts/index.js";
 import type { MapContext } from "@swooper/mapgen-core";
 import { readValidatedArtifact } from "@swooper/mapgen-core/authoring";
 import { createLabelRng } from "@swooper/mapgen-core/lib/rng";
 import { Value } from "typebox/value";
-
-import { artifacts as resourceSiteArtifacts } from "@mapgen/domain/resources/modules/sites/artifacts/index.js";
 import {
   type StandardDiscoveryPlacementMeasurements,
   StandardDiscoveryPlacementMeasurementsSchema,
@@ -54,42 +47,6 @@ function runStandardPlacementRecipe({
 
 function readResourceOutcomes(context: ReturnType<typeof runStandardPlacementRecipe>["context"]) {
   return readValidatedArtifact(context, resourceSiteArtifacts.resourcePlacementOutcomes);
-}
-
-class MismatchingResourceAdapter extends MockAdapter {
-  override placeResourceIntent(
-    width: number,
-    height: number,
-    intent: ResourcePlacementIntent
-  ): ResourcePlacementOutcome {
-    const outcome = super.placeResourceIntent(width, height, intent);
-    if (outcome.status !== "placed") return outcome;
-    return {
-      ...outcome,
-      status: "mismatch",
-      reason: "wrong-resource-type",
-      observedResourceType: outcome.resourceType + 1,
-    };
-  }
-}
-
-class UntypedResourceRejectionAdapter extends MockAdapter {
-  override placeResourceIntent(
-    width: number,
-    height: number,
-    intent: ResourcePlacementIntent
-  ): ResourcePlacementOutcome {
-    const outcome = super.placeResourceIntent(width, height, intent);
-    if (outcome.status !== "placed") return outcome;
-    return {
-      status: "rejected",
-      plotIndex: outcome.plotIndex,
-      x: outcome.x,
-      y: outcome.y,
-      resourceType: outcome.resourceType,
-      observedResourceType: outcome.observedResourceType,
-    } as ResourcePlacementOutcome;
-  }
 }
 
 class PartiallyAcceptingDiscoveryAdapter extends MockAdapter {
@@ -150,77 +107,5 @@ describe("placement reconciliation", () => {
       placedCount: 5,
       rejectedCount: 2,
     });
-  });
-
-  it("records typed resource rejections without relocation when the engine oracle rejects every intent", () => {
-    // Plan-authority cutover (S3/D4): the plan is built within static policy
-    // legality; a divergent live oracle produces typed per-intent rejections
-    // (recorded shortfalls), never relocation, type re-decision, or official
-    // generator fallback.
-    const { adapter, context } = runStandardPlacementRecipe({
-      createAdapter: ({ preset, mapInfo, mapSeed }) =>
-        createMockAdapter({
-          ...preset.dimensions,
-          mapInfo,
-          mapSizeId: preset.id,
-          rng: createLabelRng(mapSeed),
-          canHaveResource: () => false,
-        }),
-    });
-    const resourceOutcomes = readResourceOutcomes(context);
-    expect(adapter.calls.generateOfficialResources.length).toBe(0);
-    expect(adapter.calls.setResourceType.length).toBe(0);
-    expect(resourceOutcomes.summary.plannedCount).toBeGreaterThan(0);
-    expect(resourceOutcomes.summary.placedCount).toBe(0);
-    expect(resourceOutcomes.summary.rejectedCount).toBe(resourceOutcomes.summary.plannedCount);
-    expect(resourceOutcomes.summary.mismatchCount).toBe(0);
-    expect(
-      resourceOutcomes.summary.byReason.every((row) => row.reason === "cannot-have-resource")
-    ).toBe(true);
-    expect(resourceOutcomes.reconciliation.shortfalls.length).toBeGreaterThan(0);
-  });
-
-  it("fails hard when resource readback contradicts typed intent", () => {
-    let adapter: MismatchingResourceAdapter | undefined;
-
-    expect(() =>
-      runStandardPlacementRecipe({
-        createAdapter: ({ preset, mapInfo, mapSeed }) => {
-          adapter = new MismatchingResourceAdapter({
-            ...preset.dimensions,
-            mapInfo,
-            mapSizeId: preset.id,
-            rng: createLabelRng(mapSeed),
-          });
-          return adapter;
-        },
-      })
-    ).toThrow(/placement\.resources failed/i);
-    if (!adapter) throw new Error("Expected the mismatching adapter to be constructed.");
-    expect(adapter.calls.generateOfficialResources.length).toBe(0);
-    expect(adapter.calls.recalculateFertility).toBe(0);
-    expect(adapter.calls.assignAdvancedStartRegions).toBe(0);
-  });
-
-  it("fails hard when resource outcomes omit typed rejection reasons", () => {
-    let adapter: UntypedResourceRejectionAdapter | undefined;
-
-    expect(() =>
-      runStandardPlacementRecipe({
-        createAdapter: ({ preset, mapInfo, mapSeed }) => {
-          adapter = new UntypedResourceRejectionAdapter({
-            ...preset.dimensions,
-            mapInfo,
-            mapSizeId: preset.id,
-            rng: createLabelRng(mapSeed),
-          });
-          return adapter;
-        },
-      })
-    ).toThrow(/untyped rejection reason/i);
-    if (!adapter) throw new Error("Expected the rejecting adapter to be constructed.");
-    expect(adapter.calls.generateOfficialResources.length).toBe(0);
-    expect(adapter.calls.recalculateFertility).toBe(0);
-    expect(adapter.calls.assignAdvancedStartRegions).toBe(0);
   });
 });
